requirements.txt
antiorm==1.2.1
appdirs==1.4.3
apt-xapian-index==0.49
argon2-cffi==20.1.0
asgiref==3.2.10
async-generator==1.10
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
black==19.10b0
bleach==3.2.1
blinker==1.4
buildozer==1.1.1.dev0
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cmake==3.17.3
colorama==0.4.3
command-not-found==0.3
confusable-homoglyphs==3.2.0
crccheck==1.0
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.18
db==0.1.1
db-sqlite3==0.0.1
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
# Editable install with no version control (Django==3.0.5)
-e /home/sai_krupa_reddy/.local/lib/python3.8/site-packages
django-braces==1.14.0
django-crispy-forms==1.10.0
django-default-imagefield==0.5.5
django-extensions==2.2.8
django-filter==2.3.0
django-import-export==2.3.0
django-model-utils==3.1.1
django-pandas==0.6.2
django-ratings==0.3.7
django-registration==2.4.1
django-registration-redux==2.9
django-rest-framework-mongoengine==3.4.1
django-star-ratings==0.9.1
django-utils-six==2.0
djongo==1.3.4
dlib==19.20.0
entrypoints==0.3
et-xmlfile==1.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gpg===1.13.1-unknown
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
imutils==0.5.3
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-pygments==0.1.1
jupyterlab-server==1.2.0
kazam==1.4.5
keyring==18.0.1
keyrings.alt==3.4.0
Kivy==1.10.1
Kivy-Garden==0.1.4
kivymd==0.104.1
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.5.0
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
mongoengine==0.22.1
more-itertools==4.2.0
MouseInfo==0.1.3
mysql-connector-python==8.0.20
nbclient==0.5.0
nbconvert==6.0.3
nbformat==5.0.7
nest-asyncio==1.4.0
netifaces==0.10.4
nltk==3.5
notebook==6.1.4
numpy==1.18.1
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
opencv-python==4.3.0.36
openpyxl==3.0.3
packaging==20.4
pandas==1.0.1
pandocfilters==1.4.2
parso==0.7.1
pathspec==0.8.0
pep517==0.6.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
psutil==5.7.0
psycopg2==2.8.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
PyAutoGUI==0.9.50
pycairo==1.16.2
pycparser==2.20
pycrc==0.9.2
pycrypto==2.6.1
pycups==1.9.73
pygame==1.9.6
PyGetWindow==0.0.8
Pygments==2.7.1
PyGObject==3.36.0
pyjokes==0.6.0
PyJWT==1.7.1
pymongo==3.11.3
PyMsgBox==1.0.8
pyparsing==2.4.7
pyperclip==1.8.0
PyQt5==5.14.1
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.26
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-docx==0.8.10
python3-xlib==0.15
pythoncrc==1.21
pytoml==0.1.21
pyttsx3==2.88
PyTweening==1.0.3
pytz==2018.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.2
regex==2020.6.8
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.5.1
seaborn==0.10.1
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
sh==1.13.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sklearn==0.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.2.4
swapper==1.1.2.post1
systemd-python==234
tablib==2.0.0
terminado==0.9.0
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.47.0
traitlets==5.0.4
typed-ast==1.4.1
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.21
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.0.1
wikipedia==1.4.0
xlrd==1.2.0
xlwt==1.3.0
youtube-dl==2020.3.24
zipp==1.0.0
